Move DynamoDB's AttributeValue to its own module
